adding changes from dance review
